+++ src/aframe/component/architecture-toolkit/polyfloor.js
@@ -7,6 +7,7 @@ import generatePolygonBuffer from '../../../utils/data3d/buffer/get-polygon'
 import generateExtrusionBuffer from '../../../utils/data3d/buffer/get-extrusion'
 import generateNormals from '../../../utils/data3d/buffer/get-normals'
 import cloneDeep from 'lodash/cloneDeep'
+import polyfloorData3d from '../../../scene/structure/parametric-objects/polyfloor'
 
 export default {
 
@@ -16,7 +17,7 @@ export default {
 
   updateSchema: updateSchema,
 
-  update: function (oldData) {
+  update: async function (oldData) {
     var this_ = this
     var data = this_.data
 
@@ -24,19 +25,7 @@ export default {
     this.remove()
 
     // get defaults and
-    this.attributes = cloneDeep(data)
-
-    // get meshes and materials from el3d modules
-    var meshes = this.generateMeshes3d()
-
-    // clean up empty meshes to prevent errors
-    var meshKeys = Object.keys(meshes)
-    meshKeys.forEach(key => {
-      if (!meshes[key].positions || !meshes[key].positions.length) {
-        // console.warn('no vertices for mesh', key)
-        delete meshes[key]
-      }
-    })
+    let attributes = cloneDeep(data)
 
     // setup materials
     // defaults
@@ -60,12 +49,10 @@ export default {
     Object.keys(materials).forEach(mat => {
       materials[mat] = getMaterial(materials[mat])
     })
+    attributes.materials = materials;
 
     // construct data3d object
-    var data3d = {
-      meshes: meshes,
-      materials: materials
-    }
+    var data3d = await polyfloorData3d(attributes)
 
     // create new one
     this_.mesh = new THREE.Object3D()
@@ -86,81 +73,6 @@ export default {
     if (this.mesh) {
       this.el.removeObject3D('mesh')
       this.mesh = null
-    }
-  },
-
-  generateMeshes3d: function () {
-    var a = this.attributes
-
-    // a polygon can not have less than 3 points
-    if (a.polygon.length < 3) {
-      if (this.model) {
-        this.model.a.parent.remove(this.model)
-      }
-      return Promise.resolve({
-        meshes: {}
-      })
-    }
-
-    // prepare format
-    var vertices = []
-    for (var i = 0, l = a.polygon.length; i < l; i++) {
-      vertices[ i * 2 ] = a.polygon[ i ][ 0 ]
-      vertices[ i * 2 + 1 ] = a.polygon[ i ][ 1 ]
-    }
-
-    // top polygon
-    var topPolygon = generatePolygonBuffer({
-      outline: vertices,
-      y: 0,
-      uvx: a.x,
-      uvz: a.z
-    })
-
-    // ceiling polygon
-    var ceilingPolygon
-    if (a.hasCeiling) {
-      ceilingPolygon = generatePolygonBuffer({
-        outline: vertices,
-        y: a.hCeiling,
-        uvx: a.x,
-        uvz: a.z,
-        flipSide: true
-      })
-    } else {
-      ceilingPolygon = {
-        vertices: new Float32Array(0),
-        uvs: new Float32Array(0)
-      }
-    }
-
-    // sides
-    var sides = generateExtrusionBuffer({
-      outline: vertices,
-      y: -a.h,
-      flipSide: true
-    })
-
-    // return meshes
-    return {
-      top: {
-        positions: topPolygon.vertices,
-        normals: generateNormals.flat(topPolygon.vertices),
-        uvs: topPolygon.uvs,
-        material: 'top'
-      },
-      sides: {
-        positions: sides.vertices,
-        normals: generateNormals.flat(sides.vertices),
-        uvs: sides.uvs,
-        material: 'side'
-      },
-      ceiling: {
-        positions: ceilingPolygon.vertices,
-        normals: generateNormals.flat(ceilingPolygon.vertices),
-        uvs: ceilingPolygon.uvs,
-        material: 'ceiling'
-      }
     }
   }
 }
